navigation refactor
