+++ frontend/navigation/BottomNavigator.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import {BottomTabNavigationOptions, createBottomTabNavigator} from '@react-navigation/bottom-tabs';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 import HomeScreen from "@/screens/HomeScreen";
-import SearchScreen from "@/screens/SearchScreen";
 import UpdatesScreen from "@/screens/UpdatesScreen";
 import ProfileScreen from "@/screens/ProfileScreen";
 import SettingsScreen from "@/screens/SettingsScreen";
